Fix filesize overflow in graphene
Summary:
Files with sizes bigger than signed 32 bit numbers (2G) resulted in a graphene Int parsing error. To fix this we should use Decimal which is a virtually unlimited type.

Fix filesize overflow in graphene

Test Plan: I tested this locally by using a 110 Petabyte number in a Decimal, and the correct filesize was displayed in the UI without an error.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D531

+++ src/bilby/schema.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 import graphene
 from django_filters import FilterSet, OrderingFilter
 from graphene import relay
@@ -228,7 +230,7 @@ class UserDetails(graphene.ObjectType):
 class BilbyResultFile(graphene.ObjectType):
     path = graphene.String()
     is_dir = graphene.Boolean()
-    file_size = graphene.Int()
+    file_size = graphene.Decimal()
     download_id = graphene.String()
 
 
@@ -340,7 +342,7 @@ class Query(object):
                 BilbyResultFile(
                     path=f["path"],
                     is_dir=f["isDir"],
-                    file_size=f["fileSize"],
+                    file_size=Decimal(f["fileSize"]),
                     download_id=download_id
                 )
             )
